将Mock类的命名空间改为tesing（与googletmock的相同）。

+++ GoogleUnitTest/DayLineTest.cpp
@@ -168,9 +168,9 @@ namespace StockAnalysisTest {
 
   TEST_F(CStockDayLineTest, TestGetRSLogarithm) {
     CDayLine dl;
-    EXPECT_DOUBLE_EQ(dl.GetRSLogarithm(), 0.0);
+    EXPECT_DOUBLE_EQ((double)dl.GetRSLogarithm(), 0.0);
     dl.SetRSLogarithm(50.50);
-    EXPECT_DOUBLE_EQ(dl.GetRSLogarithm(), (long double)50.50);
+    EXPECT_DOUBLE_EQ((double)dl.GetRSLogarithm(), 50.50);
   }
 
   TEST_F(CStockDayLineTest, TestGetTransactionNumber) {
